Remove legacy session statistics components and logic from the SessionStats module. This includes the EngagementsStep, ParticipantsStep, ReviewStep, SessionSetupStep, SubmitStep, TargetsStep, and associated logic, streamlining the codebase for the new SessionStatsFull structure. Update documentation to reflect the changes in component organization and functionality.
