+++ src/components/SessionStatsFull/README.md
