+++ src/components/SessionStatsFull/index.ts
@@ -1,0 +1,3 @@
+export { useSessionStats } from "./useSessionStats";
+export * from "./sections";
+export * from "./types";
