+++ src/components/SessionStatsFull/sections/ParticipantsSection.tsx
