+++ src/components/SessionStatsFull/sections/ProgressIndicator.tsx
