+++ src/components/SessionStatsFull/sections/SessionConfigSection.tsx
